+++ frontend/src/components/Index/IncidentsBlock.vue
@@ -1,12 +1,15 @@
 <template>
-    <div class="row mt-4">
-        <div v-for="(incident, i) in incidents" class="col-12">
-            <h6><span class="badge badge-secondary">New</span>
-                {{incident.title}}
-            <span class="font-2">{{incident.created_at}}</span>
-            </h6>
+    <div class="row">
+        <div v-for="(incident, i) in incidents" class="col-12 mt-4">
+            <h5>Incident: {{incident.title}}<span class="font-2 float-right">{{niceDate(incident.created_at)}}</span></h5>
             {{incident.description}}
-            {{incident}}
+            <div class="row">
+                <div v-for="(update, i) in incident.updates.reverse()" v-bind:key="update.id" class="col-12 mt-3">
+                    <span class="col-2 badge text-uppercase" :class="badgeClass(update.type)">{{update.type}}</span>
+                    <span class="col-10">{{update.message}}</span>
+                    <span class="col-12 font-1 float-right text-black-50">{{ago(update.created_at)}} ago</span>
+                </div>
+            </div>
         </div>
     </div>
 </template>
@@ -30,10 +33,17 @@ export default {
         this.getIncidents()
     },
     methods: {
-        async getIncidents() {
-            this.incidents = await Api.incidents_service(this.service)
+        badgeClass(val) {
+          switch (val.toLowerCase()) {
+            case "resolved":
+              return "badge-success"
+            case "update":
+              return "badge-info"
+            case "investigating":
+              return "badge-danger"
+          }
         },
-        async getIncidentsUpdates() {
+        async getIncidents() {
             this.incidents = await Api.incidents_service(this.service)
         }
     }