+++ frontend/src/forms/CoreSettings.vue
@@ -50,37 +50,6 @@
             </div>
         </div>
 
-        <div class="row d-none">
-        <div class="col-12">
-        <h4 class="mt-5">Github Authentication</h4>
-
-        <div class="form-group row d-none">
-            <label class="col-sm-4 col-form-label">Github Client ID</label>
-            <div class="col-sm-8">
-                <input v-model="core.github_clientId" type="text" class="form-control" placeholder="" required>
-            </div>
-        </div>
-        <div class="form-group row">
-            <label class="col-sm-4 col-form-label">Github Client Secret</label>
-            <div class="col-sm-8">
-                <input v-model="core.github_clientScret" type="text" class="form-control" placeholder="" required>
-            </div>
-        </div>
-        <div class="form-group row">
-            <label for="switch-group-public" class="col-sm-4 col-form-label">Enabled</label>
-            <div class="col-md-8 col-xs-12 mt-1">
-            <span class="switch float-left">
-                <input type="checkbox" class="switch" id="switch-group-public">
-                <label for="switch-group-public">Enabled Github Auth</label>
-            </span>
-            </div>
-        </div>
-
-            <button @click.prevent="saveSettings" type="submit" class="btn btn-primary btn-block">Save Settings</button>
-
-        </div>
-        </div>
-
     </form>
 </template>
 
@@ -98,12 +67,7 @@
       methods: {
           async saveSettings() {
               const c = this.core
-              const coreForm = {
-                  name: c.name, description: c.description, domain: c.domain,
-                  timezone: c.timezone, using_cdn: c.using_cdn, footer: c.footer, update_notify: c.update_notify,
-                  gh_client_id: c.github_clientId, gh_client_secret: c.github_clientSecret
-              };
-              await Api.core_save(coreForm)
+              await Api.core_save(c)
               const core = await Api.core()
               this.$store.commit('setCore', core)
               this.core = core