+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.90.23
 - Added Incident Reporting
 - Added Cypress tests
+- Added Github and Google OAuth login (beta)
 
 # 0.90.22
 - Added range input types for integer form fields